fix(#147): 내가 응원한 가게 쿼리 옵션 파일 구조 변경 및 import 구조 재정리
