+++ src/app/(home)/_api/shop/shop.queries.ts
@@ -1,7 +1,5 @@
 import { queryOptions } from "@tanstack/react-query";
 
-import { getCheeredMember } from "@/app/(store)/_api/shop";
-
 import { getStores } from "./shop.api";
 
 export const storesQueryKeys = {
@@ -20,14 +18,4 @@ export const storesQueryOptions = ({
   queryOptions({
     queryKey: storesQueryKeys.size(size, category),
     queryFn: () => getStores({ size, category }),
-  });
-
-export const cheeredMemberQueryKeys = {
-  all: ["cheeredMember"] as const,
-};
-
-export const cheeredMemberQueryOptions = () =>
-  queryOptions({
-    queryKey: cheeredMemberQueryKeys.all,
-    queryFn: () => getCheeredMember(),
   });
